--- app/src/main/java/com/sopt/now/data/KeyStorage.kt
@@ -1,6 +1,0 @@
-package com.sopt.now.data
-
-object KeyStorage {
-    const val USER_PROFILE = "user"
-    const val USER_ID = "userId"
-}
