+++ app/src/main/java/com/sopt/now/util/KeyStorage.kt
@@ -1,0 +1,13 @@
+package com.sopt.now.data
+
+object KeyStorage {
+    const val USER_PREF = "user_pref"
+    const val USER_ID = "userId"
+    const val MEMBER = "member"
+    const val MEMBER_ID = "memberId"
+    const val JOIN = "join"
+    const val LOGIN = "login"
+    const val INFO = "info"
+    const val USERS = "users"
+    const val PAGE = "page"
+}
